feat: uses NERDTree keymap to copy file path vs let @+=expand("%")
Instead of using `<leader>yf` or whatever I come up with, we can build on
the 'yank semantics' in a NERDTree buffer.

Simply press 'y' while in NERDTree to copy the path. Or, if you're in a
current file, you can issue the following command:

  :NERDTreeFind | :normal y

The updatetime does feel a bit slow, however. This may need to be
tweaked in subsequent iterations.