feat: makes :NERDTreeFind easier, :NERDTreeToggle less
It's more often the case that I want to make local updates vs.
toggling the top-level project dir. Still, it is nice to be able to
close it.

<C-w><C-\> lets me do that while building on to the whole <C-w>(...) family
of keymaps. I think of the 'w' in <C-w> as 'window' operations. :NERDTreeToggle fits
that bill better than :NERDTreeFind which has a memoized state.