feat: replaces 'z' with 'fasd'
'fasd' is a fork of 'z' with better configuration and features.

I'm looking for a tool that will allow me to have more control over managing the frecency
list. This way, I can experiment with a 'frecency file explorer'. 'fasd' lets me
add opened buffers to the list via `fasd -A /path/to/file`.

On the other hand, 'z' only works with dirs... a severe limitation.